Add log mailing
This commit adds the ability to email the currently displayed log to a pre-set
email address. A custom "sendmail" library is used for the functionality.
To make this possible, a plaintext log rendering option was created in addition
to the HTML fragment option. An in-page pop-up "window" is displayed after the
emailing, indicating whether or not the mailing was successful.