Add data directory lookup logic
minipos will now search in a list of directories until it finds address.list
The first directory containing it will be the data directory for the session.
The directory in which the executable is found will be the library directory.
Files will be looked for in both of these in turn.

Additionally, some of the exception handling code was made more specific, to
allow unexpected exceptions to fall through and be reported in the logs.

Coincidentally, this fixes #1.