Remove the "confirmed only" parameter in get_balance()
This change is API-breaking.
With all explorers now providing both the confirmed and unconfirmed balance, it
makes no sense to continue supporting the partial values anymore.
Truth be told, i got sick and tired of trying to work around the quirks of this
or that data provider, especially in ways that significantly alter the workings
of my other code.
I will look back at BlockChair only when it starts providing all the required
information in one API call, like the rest of the players do.