Add double spend detection to backend and frontend
When a double spend is detected, MiniPOS will switch to a "waiting for
confirmation" mode: explorers are polled more rarely, and the only variable
checked is the number of confirmations a transaction has. The operator can
cancel the wait at any time, as normal, and a prominent warning is displayed.

Additionally, the "payment received" popup was modified to match the "double
spend detected" popup