Enforce stricter template replacement policy
This is a BREAKING CHANGE; Any modifications made to existing installations
must be reapplied manually. A summary of changes:

* All HTML templates were moved to the "templates" directory;
* Only the "welcome_footer", "log_header", "log_footer" and "invoice_text"
templates can be overriden by placing them in the data directory (in the
"templates" subdirectory);
* All assets, except "scripts.js" can be overriden as before;
* The Bitcoin Cash logo and icon were given more generic names.

The main effect from this change is that the user running MiniPOS cannot
override its core functionality when it is installed as a different user, or as
root.